1.删除回款计划接口，编辑回款计划接口，调通 2.回款计划整体流程完成

+++ app/src/main/java/com/loyo/oa/v2/activityui/order/OrderAddActivity.java
@@ -190,7 +190,7 @@ public class OrderAddActivity extends BaseActivity implements View.OnClickListen
         map.put("customerName",customerName);
         map.put("title",et_name.getText().toString());
         map.put("attachmentUUId","");
-        map.put("dealMoney",Integer.parseInt(et_money.getText().toString()));
+        map.put("dealMoney",Float.parseFloat(et_money.getText().toString()));
         map.put("orderNum",et_ordernum.getText().toString());
         map.put("remark",et_remake.getText().toString());
         map.put("proInfo",productData);
